% Generated by roxygen2 (4.0.1): do not edit by hand
\name{addErrorMessage}
\alias{addErrorMessage}
\title{Adds an error message to error message list}
\usage{
addErrorMessage(text)
}
\description{
Called from the plotTests function and not meant to be
called by user.
}
\details{
This function adds an error message to the list of error messages
which is passed to RExcel. The messages are shown to the user
in the Excel spreadsheet.
}

