% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsMixedModel.R
\name{calculateDaysSinceInclusion}
\alias{calculateDaysSinceInclusion}
\title{Helper function for calculating the number of days from first date}
\usage{
calculateDaysSinceInclusion(data, subjectIDVar, dateVar)
}
\arguments{
\item{data}{The data.}

\item{subjectIDVar}{Which variable identifies the subject.}

\item{dateVar}{Whic variable contains the date.}
}
\description{
Calculates and returns the number of days since minimum date of a certain subject.
Adds the number of days since min date as column to the data frame it returns.
}

