% Generated by roxygen2 (4.0.1): do not edit by hand
\name{checkifCombinationExists}
\alias{checkifCombinationExists}
\title{Checks if combination of groups exists}
\usage{
checkifCombinationExists(data, firstLevel, secondLevel)
}
\description{
Checks if combination of groups exists and is not meant
to be called from by user.
}
\details{
Checks for existance of combination of first and second level groups
in the data.
}

