% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsDetermineTypeofVariable.R
\name{determineTypeofVariable}
\alias{determineTypeofVariable}
\title{Determine the type of variable}
\usage{
determineTypeofVariable(variableValues)
}
\arguments{
\item{variableValues}{takes a vector of values}
}
\value{
Returns the number of levels and type of variable.
}
\description{
Determines what type of variable is used. Used in regression modeling.
}

