% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsGenerateGraph.R
\name{downloadPlot}
\alias{downloadPlot}
\title{Code to generate a downloadable file containing plot (downloaHandler) to download using a button.}
\usage{
downloadPlot(filename = "plot.eps", plotFunction, width, height,
  print = FALSE)
}
\arguments{
\item{filename}{Suggested file name.}

\item{plotFunction}{The function that plots the graph (can be reactive).}

\item{width}{Width of the plot in pixels.}

\item{height}{Height of plot in pixels.}

\item{print}{Whether to use print() function or not. Needed with ggplot objects and not with base plots?}
}
\description{
Code to generate a downloadable file containing plot (downloaHandler) to download using a button.
}

