\name{drawLevelLabels}
\alias{drawLevelLabels}
\title{Draw labels and lines for groups}
\usage{
drawLevelLabels(data, TYPE.LEVELS, LINE.SIZE, DIAGNOSIS.LEVELS, firstLevel,
  secondLevel)
}
\description{
Function that plots labels for both possible groups
describing subjects.
}
\details{
The function is meant to be called by the general function
for plotting, when it needs to draw the labels for the
groups.
}

