% Generated by roxygen2 (4.0.2): do not edit by hand
\name{drawLineBelow2ndLevel}
\alias{drawLineBelow2ndLevel}
\title{Draws lines below 2nd level groups}
\usage{
drawLineBelow2ndLevel(data, firstLevel, secondLevel, lastRowforDiagnosis,
  DIAGNOSIS.LEVELS, LINE.SIZE)
}
\description{
Not meant to be called by the user.
}

