% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsGenerateDescriptions.R
\name{generateDescription}
\alias{generateDescription}
\title{Generate descriptions of outputs}
\usage{
generateDescription(outputName)
}
\arguments{
\item{outputName}{Takes the output name (label) as argument.}
}
\description{
Generates descriptions for output (tables, graphs).
}

