% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsImportData.R
\name{importSymptomsData}
\alias{importSymptomsData}
\title{Function that imports data for the plotSymptoms plot}
\usage{
importSymptomsData(datafile, format)
}
\arguments{
\item{datafile}{Path to the file containing data.}

\item{format}{Format of the file containing data ("Excel", "TSV", "Demo").}
}
\description{
Function is called to import the data in appropriate format for the
plotSymptoms plot.
}

