% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsImportData.R
\name{importSymptomsPatients}
\alias{importSymptomsPatients}
\title{Function that imports PATIENTS Excel tab for the plotSymptoms plot}
\usage{
importSymptomsPatients(datafile)
}
\arguments{
\item{datafile}{Path to the file containing data.}
}
\description{
Function that imports PATIENTS Excel tab for the plotSymptoms plot
}

