\name{isolateTests}
\alias{isolateTests}
\title{Gets all test results written in a cell}
\usage{
isolateTests(string, separator)
}
\description{
Not meant to be called by the user.
}
\details{
Parses test result strings from a string (a cell of data
containing results). Different separators can be defined.
Leading spaces are removed.
}

