% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_GroupingVariable.R
\name{MannWhitneyPValue}
\alias{MannWhitneyPValue}
\title{Helper function that returns a P values calculated by Mann Whitney}
\usage{
MannWhitneyPValue(group1, group2)
}
\arguments{
\item{group1}{Values for group 1.}

\item{group2}{Values for group 2}
}
\description{
Helper function that returns a P values calculated by Mann Whitney
}

