% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/RunApps.R
\name{medplotOnline}
\alias{medplotOnline}
\title{Function to load the medplot plotSymptoms shiny web app}
\usage{
medplotOnline()
}
\description{
Function call loads the shiny app in located the default installation folder.
}

