% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsMixedModel.R
\name{mixedModel}
\alias{mixedModel}
\title{Return a list of result for mixed modeling}
\usage{
mixedModel(data, selectedSymptoms, coVariate1st, subjectIDVar, measurementVar,
  dateVar, thresholdValue, treatasBinary, selectedModel)
}
\arguments{
\item{data}{The data.}

\item{selectedSymptoms}{Which outcome variables were selected.}

\item{coVariate1st}{Which covariate was selected.}

\item{subjectIDVar}{Which variable identifies the subject.}

\item{measurementVar}{Which variable represent the measurement occasion.}

\item{dateVar}{Which variable represents the date.}

\item{thresholdValue}{What was the selected theshold value.}

\item{treatasBinary}{Should the data be treated as binary.}

\item{selectedModel}{Which modeling variant was selected.}
}
\description{
Returns a list containing results of mixed modeling.
The list contains the printable (for GUI output) and raw results (for ploting).
}

