% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsLasagnaPlot.R
\name{pastePlotFilenames}
\alias{pastePlotFilenames}
\title{Paste a list of filenames (helper function for lasagna plots)}
\usage{
pastePlotFilenames(filenames)
}
\arguments{
\item{filenames}{The list of filenames that were the output of lasagna plots}
}
\value{
Returns a vector containing paths to generated files (plots).
}
\description{
Paste a list of filenames (helper function for lasagna plots)
}

