\name{plot.rcs.mod}
\alias{plot.rcs.mod}
\title{Plot RSC MOD ??? TODO : help contents}
\usage{
\method{plot}{rcs.mod}(My.mod, my.var, my.ylab = "Response",
  my.xlab = "Covariate", my.xlim = NULL, my.ylim = NULL,
  my.knots = NULL, my.title = NULL, vlines = NULL, hlines = NULL,
  my.cex.lab = NULL, my.axes = TRUE, my.cex.main = NULL,
  my.cex.axis = NULL)
}
\arguments{
  \item{My.mod}{???}
}
\description{
Plot RSC MOD ??? TODO : help contents
}

