% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_MeasurementOccasion.R
\name{plotCI}
\alias{plotCI}
\title{Ploting confidence intervals}
\usage{
plotCI(data.yn, measurements, selectedSymptoms, selectedProportion)
}
\arguments{
\item{ddd}{dddd}
}
\description{
Ploting confidence intervals
}

