\name{plotClusterDendrogram}
\alias{plotClusterDendrogram}
\title{Plot dendrogram of patient groups based on symptoms.}
\usage{
plotClusterDendrogram(data, variableName, variableValue, selectedSymptoms)
}
\arguments{
  \item{data}{Data fram to be passed to the function.}

  \item{variableName}{The column name of the variable used
  for filtering.}

  \item{variableValues}{The value of the filtering variable
  used to filter.}
}
\description{
Plots a dendrogram of patient groups based on symptom
values they have.
}

