% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_MeasurementOccasion.R
\name{plotDistribution}
\alias{plotDistribution}
\title{Plots of distribution of the symptoms}
\usage{
plotDistribution(data, selectedSymptoms, selectedProportion, measurements)
}
\arguments{
\item{data}{Data frame for ploting.}

\item{selectedSymptoms}{Column names of symptoms in data frame.}

\item{selectedProportion}{User selected measurement time.}

\item{measurements}{Vector of measurement values.}
}
\description{
Plots of distribution of the symptoms
}

