\name{plotDistributionBoxplot}
\alias{plotDistributionBoxplot}
\title{Plots boxplots of symptoms.}
\usage{
plotDistributionBoxplot(data, data.yn, selectedSymptoms, selectedProportion,
  measurements, posOnly = FALSE, threshold)
}
\arguments{
  \item{data}{Data for the boxplots.}

  \item{selectedSymptoms}{Column names of symptoms in data
  frame.}

  \item{selectedProportion}{User selected measurement
  time.}

  \item{measurements}{Vector of measurement values.}

  \item{posOnly}{TRUE/FALSE}

  \item{threshold}{User selected threshold value of
  symptoms.}
}
\description{
Plots boxplots of symptoms.
}

