% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsMixedModel.R
\name{plotFixedEffectsofcoVariate1st}
\alias{plotFixedEffectsofcoVariate1st}
\title{Plot the fixed effects for mixed models (helper function)}
\usage{
plotFixedEffectsofcoVariate1st(calculatedStatistics, coVariate1st,
  coVariate1stReferenceValue, treatasBinary, variableOrder)
}
\arguments{
\item{calculatedStatistics}{The results of modeling.}

\item{coVariate1st}{Which variable represents the covariate.}

\item{coVariate1stReferenceValue}{What is the reference level of the covariate.}

\item{treatasBinary}{Should outcomes be treated as binary.}

\item{variableOrder}{What should be the variable order on graph.}
}
\description{
Plot the fixed effects for mixed models (helper function)
}

