% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsMixedModel.R
\name{plotFixedEffectsofDaysSinceInclusion}
\alias{plotFixedEffectsofDaysSinceInclusion}
\title{Plot the fixed effects for mixed models (helper function)}
\usage{
plotFixedEffectsofDaysSinceInclusion(calculatedStatistics, treatasBinary,
  variableOrder)
}
\arguments{
\item{calculatedStatistics}{The results of modeling.}

\item{treatasBinary}{Should outcomes be treated as binary.}

\item{variableOrder}{What should be the variable order on graph.}
}
\description{
Plot the fixed effects for mixed models (helper function)
}

