% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsLasagnaPlot.R
\name{plotLasagna}
\alias{plotLasagna}
\title{Plot a lasagna plot (heatmap)}
\usage{
plotLasagna(data, treatasBinary, symptom, dateVar, personIDVar, measurementVar,
  groupingVar, thresholdValue)
}
\arguments{
\item{data}{The filtered data.}

\item{treatasBinary}{Flag if the data should be treated a binary.}

\item{symptom}{Which outcome variables are selected.}

\item{dateVar}{Which varible represents the date.}

\item{personIDVar}{Which varible represents the person ID.}

\item{measurementVar}{Which varible represents the measurement occasion.}

\item{groupingVar}{Which varible represents the grouping variable.}

\item{thresholdValue}{The value of the binary threshold.}
}
\description{
Plots a lasagna plot (a heatmap) of the data.
}

