% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsPresencePlot.R
\name{plotPresenceofSymptoms}
\alias{plotPresenceofSymptoms}
\title{Plot the presence of symptoms}
\usage{
plotPresenceofSymptoms(data, selectedSymptoms, measurementVar, measurement,
  thresholdValue = 0)
}
\arguments{
\item{data}{The data.}

\item{measurementVar}{Which variable holds evaluation occasions.}

\item{measurement}{Which evaluation occasio was chosen.}

\item{thresholdValue}{What is the threshold for dichotomizing.}

\item{Which}{outcome variables were selected.}
}
\description{
Plot the presence of symptoms
}

