% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_GroupingVariable.R
\name{plotPropPositiveCI}
\alias{plotPropPositiveCI}
\title{Plot of confidence intervals for proportions for groups an measurements.}
\usage{
plotPropPositiveCI(data, groupingVar, measurementVar, selectedSymptoms)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}

