% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabRCS.R
\name{plotRCS}
\alias{plotRCS}
\title{Plot RCS}
\usage{
plotRCS(data.all, data.yn = NULL, measurement, selectedSymptoms,
  measurementSelectedrcs, rcsIDVar, binaryVar = TRUE)
}
\arguments{
\item{data}{Data for ploting.}
}
\description{
Plot RCS.
}

