% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabRCS.R
\name{plotRCSmod}
\alias{plotRCSmod}
\title{Helper function for plotRCS()}
\usage{
plotRCSmod(My.mod, my.var, my.ylab = "Response", my.xlab = "Covariate",
  my.xlim = NULL, my.ylim = NULL, my.knots = NULL, my.title = NULL,
  vlines = NULL, hlines = NULL, my.cex.lab = NULL, my.axes = TRUE,
  my.cex.main = NULL, my.cex.axis = NULL)
}
\arguments{
\item{My.mod}{???}

\item{my.var}{}

\item{my.ylab}{}

\item{my.xlab}{}

\item{my.xlim}{}

\item{my.ylim}{}

\item{my.knots}{}

\item{my.title}{}

\item{vlines}{}

\item{hlines}{}

\item{my.cex.lab}{}

\item{my.axes}{}

\item{my.cex.main}{}

\item{my.cex.axis}{}
}
\description{
Helper function for plotRCS()
}

