\name{plotSymptomsTimeline}
\alias{plotSymptomsTimeline}
\title{Plot timeline of symptom presence}
\usage{
plotSymptomsTimeline(data)
}
\arguments{
  \item{data}{The data frame used by the plot. Data must be
  in appropriate format for ggplot() (see melt()). Returns
  a ggplot object that has to be plotted via print().}
}
\description{
Function that plots symptom severity presence for patients
at a certain time.
}

