% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_OverTime.R
\name{plotTimelineBoxplots}
\alias{plotTimelineBoxplots}
\title{Plot timeline boxplots}
\usage{
plotTimelineBoxplots(data, personIDVar, measurementVar, selectedSymptoms,
  faceting)
}
\arguments{
\item{data}{The data.}

\item{personIDVar}{Which var identifies the person.}

\item{measurementVar}{Which var identifies the evaluation occasions.}

\item{selectedSymptoms}{Which outcome variables wre selected.}

\item{faceting}{Should faceting be used.}
}
\description{
Plots boxplots for each evaluation occasion.
}

