\name{plotTimelineProfiles}
\alias{plotTimelineProfiles}
\title{Plot profiles for variables.}
\usage{
plotTimelineProfiles(data, plotType = "oneGraph", personIDVar, measurementVar,
  selectedSymptoms, sizeofRandomSample = 10, sizeofGroup = 25)
}
\arguments{
  \item{TODO}{}
}
\description{
TODO
}

