% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsValueswithCIsPlot.R
\name{plotValueswithCIs}
\alias{plotValueswithCIs}
\title{A generic plot that plots values and their confidence intervals}
\usage{
plotValueswithCIs(data, variableName = "Variable", valueName = "OR",
  CILowerName = "CILower", CIUpperName = "CIUpper", xLabel, yLabel,
  graphTitle, vLine = NULL, variableOrder = NULL)
}
\arguments{
\item{data}{The data to plot.}

\item{variableName}{Label for the variables plotted.}

\item{valueName}{Variable that holds values.}

\item{CILowerName}{Variable that holds lower CI value.}

\item{CIUpperName}{Variable that holds upper CI value.}

\item{xLabel}{Horizontal label.}

\item{yLabel}{Vertical label.}

\item{graphTitle}{Graph title.}

\item{vLine}{Where to plot a vertical line.}

\item{variableOrder}{The order of the variables on the graph.}
}
\description{
A generic plot that plots values and their confidence intervals
}

