% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_GroupingVariable.R
\name{propPValue}
\alias{propPValue}
\title{Helper function that returns a P value for proportions}
\usage{
propPValue(successG1, failureG1, successG2, failureG2)
}
\arguments{
\item{successG1}{Number of successes in group 1.}

\item{failureG1}{Number of failures in group 1.}

\item{successG2}{... in group 2.}

\item{failureG2}{... in group 2.}
}
\description{
Helper function that returns a P value for proportions
}

