% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsMixedModel.R
\name{resortbyVariables}
\alias{resortbyVariables}
\title{Helper function to resort printable data frames to keep Variable levels together}
\usage{
resortbyVariables(dataframe, selectedSymptoms)
}
\arguments{
\item{dataframe}{The data.}

\item{selectedSymptoms}{Which outcome variables were selected.}
}
\description{
Resorts the variables according the order the outcomes were selected on the GUI.
Returns the resorted dataframe.
}

