% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_GroupingVariable.R
\name{returnPropCIs}
\alias{returnPropCIs}
\title{Helper function to return proportions with CIs}
\usage{
returnPropCIs(data, groupingVar, measurementVar, selectedSymptoms)
}
\arguments{
\item{data}{The data.}

\item{groupingVar}{The grouping varible.}

\item{measurementVar}{The measurement variable.}

\item{selectedSymptoms}{The symptoms selected.}
}
\description{
Helper function to return proportions with CIs
}

