% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDataSummary.R
\name{summarizeData}
\alias{summarizeData}
\title{Tables summarizing data}
\usage{
summarizeData(data, personIDVar, measurementVar, selectedSymptoms, groupingVar)
}
\arguments{
\item{data}{The data.}

\item{personIDVar}{Which variable identifies the person.}

\item{measurementVar}{Which variable identifies the evaluation occasion.}

\item{selectedSymptoms}{Which outcome variables were selected.}

\item{groupingVar}{Which variable is used for grouping.}
}
\description{
Outputs a summary of the data.
}

