% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_OverTime.R
\name{tabelizeBoxplots}
\alias{tabelizeBoxplots}
\title{Boxplots data in form of a table}
\usage{
tabelizeBoxplots(measurements, measurementVar, data, selectedSymptoms)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}

