\name{tabelizeLogistf}
\alias{tabelizeLogistf}
\title{OBSOLETE Plot logistf
#
#}
\usage{
tabelizeLogistf(data, measurementVar, selectedMeasurement, covariate,
  selectedSymptoms, thresholdValue)
}
\arguments{
  \item{TODO}{TODO write instructions}

  \item{TODO}{}
}
\description{
TODO write description # #
}
\details{
#
}

