% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_MeasurementOccasion.R
\name{tableAllWithSymptoms}
\alias{tableAllWithSymptoms}
\title{Table with descritpive statistics for both groups}
\usage{
tableAllWithSymptoms(data, measurementVar, forMeasurement, symptomsNames,
  thresholdValue = 0)
}
\arguments{
\item{data}{The data}

\item{measurementVar}{The mesurement variable.}

\item{forMeasurement}{For which measurement occasion.}

\item{symptomsNames}{For which outcomes.}

\item{thresholdValue}{Threshold for dichotomizing.}
}
\description{
Thresholds taken into account for proportions, not for medians.
}

