% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsLinearModel.R
\name{tableLinear}
\alias{tableLinear}
\title{Return a list of result for linear modeling}
\usage{
tableLinear(data, measurementVar, selectedMeasurement, covariate,
  selectedSymptoms)
}
\arguments{
\item{data}{The data.}

\item{measurementVar}{Which variable represent the measurement occasion.}

\item{selectedMeasurement}{Which measurement occasion was selected.}

\item{Which}{covariates are to be used.}

\item{Which}{outcome variables were selected.}
}
\description{
Returns a list containing results of linear modeling.
The list contains the printable (for GUI output) and raw results (for ploting).
}

