% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsLogisticRegression.R
\name{tableLogist}
\alias{tableLogist}
\title{Return a list of result for logistic regression modeling}
\usage{
tableLogist(data, measurementVar, selectedMeasurement, covariate,
  selectedSymptoms, thresholdValue)
}
\arguments{
\item{data}{The data.}

\item{measurementVar}{Which variable represent the measurement occasion.}

\item{selectedMeasurement}{Which measurement occasion was selected.}

\item{covariate}{Which covariates are to be used.}

\item{selectedSymptoms}{Which outcome variables were selected.}

\item{thresholdValue}{The threshold value for dichotomizing.}
}
\description{
Returns a list containing results of logistic regression modeling.
The list contains the printable (for GUI output) and raw results (for ploting).
}

