% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_OverTime.R
\name{tableMedians}
\alias{tableMedians}
\title{Helper function for tabelizeBoxplots}
\usage{
tableMedians(measurement, measurementVar, data, selectedSymptoms)
}
\arguments{
\item{measurement}{Selected evaluation occasion.}

\item{measurementVar}{Which var identifies the evaluation occasions.}

\item{data}{The data.}

\item{selectedSymptoms}{Which outcome variables were selected.}
}
\description{
Helper function for tabelizeBoxplots
}

