% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_OverTime.R
\name{tableProportions}
\alias{tableProportions}
\title{Construct a table of proportions}
\usage{
tableProportions(measurement, measurementVar, data, selectedSymptoms)
}
\arguments{
\item{measurement}{Selected evaluation occasion.}

\item{measurementVar}{Which var identifies the evaluation occasions.}

\item{data}{The data.}

\item{selectedSymptoms}{Which outcome variables were selected.}
}
\description{
Construct a table of proportions
}

