% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_GroupingVariable.R
\name{tablePropPosGroups}
\alias{tablePropPosGroups}
\title{Table with proportions}
\usage{
tablePropPosGroups(data, groupingVar = "Sex", measurementVar, forMeasurement,
  symptomsNames, thresholdValue = 0, doPValueAdjustments)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}

