% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tableRCS}
\alias{tableRCS}
\title{Restricted cubic spline P values in tabular form}
\usage{
tableRCS(data.all, data.yn, measurement, selectedSymptoms,
  measurementSelectedrcs, rcsIDVar, binaryVar = TRUE)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}

