\name{importSymptomsPatients}
\alias{importSymptomsPatients}
\title{Function that imports PATIENTS Excel tab for the plotSymptoms plot}
\usage{
importSymptomsPatients(datafile)
}
\arguments{
  \item{datafile}{Path to the file containing data.}
}
\description{
Function that imports PATIENTS Excel tab for the
plotSymptoms plot
}

