% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_GroupingVariable.R
\name{permCorrPValue}
\alias{permCorrPValue}
\title{Multivariate permutation test for correcting the P value}
\usage{
permCorrPValue(data, nPermutations = 999, selectedSymptoms, groupingVar,
  measurementVar, measurementOccasion, FUN, thresholdValue = 0)
}
\arguments{
\item{data}{The data.}

\item{nPermutations}{Number of permutations.}

\item{selectedSymptoms}{Which outcomes were selected.}

\item{groupingVar}{The grouping variable.}

\item{measurementVar}{The evaluation occasion variable.}

\item{measurementOccasion}{For which evaluation occasion.}

\item{thresholdValue}{Dichotomizing threshold.}

\item{Which}{function to use for returning a P value.}
}
\description{
Multivariate P values correction.
}

