% Generated by roxygen2 (4.0.2): do not edit by hand
\name{plotClusterHeatmap}
\alias{plotClusterHeatmap}
\title{Plot clustered heatmaps.}
\usage{
plotClusterHeatmap(data, variableName, variableValue, selectedSymptoms,
  annotationVars = NA, treatasBinary = FALSE)
}
\arguments{
\item{data}{Data fram to be passed to the function.}

\item{variableName}{The column name of the variable used for filtering.}

\item{variableValues}{The value of the filtering variable used to filter.}
}
\description{
Plots a heatmap for presence of symptoms.
}

