% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotDistribution}
\alias{plotDistribution}
\title{Plots of distribution of the symptoms}
\usage{
plotDistribution(data, selectedSymptoms, selectedProportion, measurements)
}
\arguments{
\item{data}{Data frame for ploting.}
}
\description{
Plots of distribution of the symptoms
}

