% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_GroupingVariable.R
\name{plotPropPositiveCI}
\alias{plotPropPositiveCI}
\title{Plot of confidence intervals}
\usage{
plotPropPositiveCI(data, groupingVar, measurementVar, selectedSymptoms)
}
\arguments{
\item{data}{The data.}

\item{groupingVar}{The grouping variable.}

\item{measurementVar}{Which is the measurement variable.}

\item{selectedSymptoms}{For which outcomes.}
}
\description{
Plot of confidence intervals for proportions for groups and measurements.
}

