% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotPropPositiveCI}
\alias{plotPropPositiveCI}
\title{Plot of confidence intervals for proportions for groups an measurements.}
\usage{
plotPropPositiveCI(data, groupingVar, measurementVar, selectedSymptoms)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}

