% Generated by roxygen2 (4.0.1): do not edit by hand
\name{plotRCS}
\alias{plotRCS}
\title{Plot RCS}
\usage{
plotRCS(data.all, data.yn = NULL, measurement, selectedSymptoms,
  measurementSelectedrcs, rcsIDVar, binaryVar = TRUE)
}
\arguments{
\item{data}{Data for ploting.}
}
\description{
Plot RCS.
}

