% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabRCS.R
\name{plotRCS}
\alias{plotRCS}
\title{Plot RCS}
\usage{
plotRCS(data.all, data.yn = NULL, measurement, selectedSymptoms,
  measurementSelectedrcs, rcsIDVar, binaryVar = TRUE)
}
\arguments{
\item{data.all}{Data for ploting.}

\item{data.yn}{???}

\item{measurement}{???}

\item{selectedSymptoms}{???}

\item{measurementSelectedrcs}{???}

\item{rcsIDVar}{???}

\item{binaryVar}{???}
}
\description{
Plot restricted cubic splines.
}

