\name{plotRCSmod}
\alias{plotRCSmod}
\title{Plot RCS MOD ??? TODO : help contents}
\usage{
plotRCSmod(My.mod, my.var, my.ylab = "Response", my.xlab = "Covariate",
  my.xlim = NULL, my.ylim = NULL, my.knots = NULL, my.title = NULL,
  vlines = NULL, hlines = NULL, my.cex.lab = NULL, my.axes = TRUE,
  my.cex.main = NULL, my.cex.axis = NULL)
}
\arguments{
  \item{My.mod}{???}
}
\description{
Plot RCS MOD ??? TODO : help contents
}

