% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_OverTime.R
\name{plotTimelineBoxplots}
\alias{plotTimelineBoxplots}
\title{Plot timeline boxplots}
\usage{
plotTimelineBoxplots(data, personIDVar, measurementVar, selectedSymptoms,
  faceting)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}

