% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_OverTime.R
\name{plotTimelineProfiles}
\alias{plotTimelineProfiles}
\title{Plot profiles for variables.}
\usage{
plotTimelineProfiles(data, plotType = "oneGraph", personIDVar, measurementVar,
  selectedSymptoms, sizeofRandomSample = 10, sizeofGroup = 25)
}
\arguments{
\item{data}{The data.}

\item{plotType}{What type of plot (oneGraph, randomSample, multipleGraphs).}

\item{personIDVar}{Which variable identifies the person.}

\item{measurementVar}{Which var identifies the evaluation occasion.}

\item{selectedSymptoms}{Which outcomes were selected.}

\item{sizeofRandomSample}{Size of random sample (in case of oneGraph variant).}

\item{sizeofGroup}{Size of group (in case of multipleGraphs).}
}
\description{
Plots different kind of profile (spaghetti) plots
for the outcome values over time.
}

