% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDataSummary.R
\name{summarizeData}
\alias{summarizeData}
\title{Tables summarizing data}
\usage{
summarizeData(data, personIDVar, measurementVar, selectedSymptoms, groupingVar)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}

