% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_MeasurementOccasion.R
\name{tableAllWithSymptoms}
\alias{tableAllWithSymptoms}
\title{Table for both groups}
\usage{
tableAllWithSymptoms(data, measurementVar, forMeasurement, symptomsNames,
  thresholdValue = 0)
}
\arguments{
\item{TODO}{}
}
\description{
TODO Thresholds take into account for proportions, not for medians.
}

