\name{tableAllWithSymptoms}
\alias{tableAllWithSymptoms}
\title{Table for both groups}
\usage{
tableAllWithSymptoms(data, measurementVar, forMeasurement, symptomsNames,
  thresholdValue = 0)
}
\arguments{
  \item{TODO}{}
}
\description{
TODO Thresholds take into account for proportions, not for
medians.
}

