% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabLogistf.R
\name{tableLogistf}
\alias{tableLogistf}
\title{Logistf data in tabular format}
\usage{
tableLogistf(data, measurementVar, selectedMeasurement, covariate,
  selectedSymptoms, thresholdValue)
}
\arguments{
\item{data}{The data.}

\item{measurementVar}{Which variable represents the evaluatuon occasions.}

\item{selectedMeasurement}{For which evaluation occasion.}

\item{covariate}{Which var is the covariate.}

\item{selectedSymptoms}{For which outcomes.}

\item{thresholdValue}{For which threshold value.}
}
\description{
TODO
}

