% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tableLogistf}
\alias{tableLogistf}
\title{OBSOLETE Plot logistf
#
#}
\usage{
tableLogistf(data, measurementVar, selectedMeasurement, covariate,
  selectedSymptoms, thresholdValue)
}
\arguments{
\item{TODO}{TODO write instructions}

\item{TODO}{}
}
\description{
TODO write description
#
#
}
\details{
#
}

