\name{tableMediansWithSymptoms}
\alias{tableMediansWithSymptoms}
\title{Table with medians}
\usage{
tableMediansWithSymptoms(data, groupingVar = "Sex", measurementVar,
  forMeasurement, symptomsNames, thresholdValue = 0, doPValueAdjustments)
}
\arguments{
  \item{TODO}{}
}
\description{
TODO Ignores threshold value.
}

