% Generated by roxygen2 (4.0.2): do not edit by hand
\name{tableMeGroups}
\alias{tableMeGroups}
\title{Table with medians}
\usage{
tableMeGroups(data, groupingVar = "Sex", measurementVar, forMeasurement,
  symptomsNames, thresholdValue = 0, doPValueAdjustments)
}
\arguments{
\item{TODO}{}
}
\description{
TODO Ignores threshold value.
}

