\name{tableMeGroups}
\alias{tableMeGroups}
\title{Table with medians}
\usage{
tableMeGroups(data, groupingVar = "Sex", measurementVar, forMeasurement,
  symptomsNames, thresholdValue = 0, doPValueAdjustments)
}
\arguments{
  \item{TODO}{}
}
\description{
TODO Ignores threshold value.
}

