% Generated by roxygen2 (4.0.1): do not edit by hand
\name{tablePropPosGroups}
\alias{tablePropPosGroups}
\title{Table with proportions}
\usage{
tablePropPosGroups(data, groupingVar = "Sex", measurementVar, forMeasurement,
  symptomsNames, thresholdValue = 0, doPValueAdjustments)
}
\arguments{
\item{TODO}{}
}
\description{
TODO
}

