% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabDistribution_GroupingVariable.R
\name{tablePropPosGroups}
\alias{tablePropPosGroups}
\title{Table with proportions}
\usage{
tablePropPosGroups(data, groupingVar = "Sex", measurementVar, forMeasurement,
  symptomsNames, thresholdValue = 0, doPValueAdjustments)
}
\arguments{
\item{data}{The data.}

\item{groupingVar}{Which is the grouping variable.}

\item{measurementVar}{Which is the measurement variable.}

\item{forMeasurement}{For which measuremet occasion?}

\item{symptomsNames}{For which outcome variables?}

\item{thresholdValue}{Dichotomizing threshold.}

\item{doPValueAdjustments}{Should P value adjustments be done?}
}
\description{
Table of proportions of patients that have a certain outcome present.
}

