% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/PlotSymptomsTabRCS.R
\name{tableRCS}
\alias{tableRCS}
\title{Restricted cubic spline P values in tabular form}
\usage{
tableRCS(data.all, data.yn, measurement, selectedSymptoms,
  measurementSelectedrcs, rcsIDVar, binaryVar = TRUE)
}
\arguments{
\item{data.all}{???}

\item{data.yn}{}

\item{measurement}{}

\item{selectedSymptoms}{}

\item{measurementSelectedrcs}{}

\item{rcsIDVar}{}

\item{binaryVar}{}
}
\description{
Restricted cubic spline P values in tabular form
}

